Sprint-8/Backend/Sprint_8/requirements.txt
Django==5.1.2
djangorestframework==3.15.2
django-cors-headers==4.6.0